google-examples/DBHackbot/requirements.txt
streamlit
google-cloud-aiplatform
google-cloud-discoveryengine
google-api-core
